♻️ refactor(cli): modularize release logic into separate module
- extract release-related functions into a new `release.rs` module for better code organization
- update imports and function calls to reflect the new module structure

🔥 remove(main.rs): delete unused functions for cleanup

- remove unused functions `push_committed`, `commit_changed_files`, `run_release`, `release_semver`, `release_workspace`, `release_package`, and `print_changelog` from `main.rs`
- clean up imports to remove dependencies on deleted functions
