+++ src/bin/main.rs
@@ -1,18 +1,12 @@
-use std::{env, fs, path::Path};
-
 use clap::Parser;
-use config::Config;
 use env_logger::Env;
-use owo_colors::{OwoColorize, Style};
-use pcu::{Client, Error, GitOps, MakeRelease, Sign, Workspace};
 
 use color_eyre::Result;
 
 const LOG_ENV_VAR: &str = "RUST_LOG";
 const LOG_STYLE_ENV_VAR: &str = "RUST_LOG_STYLE";
-const GITHUB_PAT: &str = "GITHUB_TOKEN";
 
-use pcu::cli::{CIExit, Cli, Commands, Release};
+use pcu::cli::{CIExit, Cli, Commands};
 use tracing_subscriber::EnvFilter;
 
 #[tokio::main]
@@ -31,7 +25,7 @@ async fn main() -> Result<()> {
         Commands::Commit(commit_args) => pcu::cli::run_commit(sign, commit_args).await,
         Commands::Push(push_args) => pcu::cli::run_push(push_args).await,
         Commands::Label(label_args) => pcu::cli::run_label(label_args).await,
-        Commands::Release(rel_args) => run_release(sign, rel_args).await,
+        Commands::Release(rel_args) => pcu::cli::run_release(sign, rel_args).await,
     };
 
     match res {
@@ -54,193 +48,6 @@ async fn main() -> Result<()> {
     Ok(())
 }
 
-async fn push_committed(
-    client: &Client,
-    prefix: &str,
-    tag_opt: Option<&str>,
-    no_push: bool,
-) -> Result<()> {
-    log::info!("Push the commit");
-    log::trace!("tag_opt: {tag_opt:?} and no_push: {no_push}");
-
-    client.push_commit(prefix, tag_opt, no_push)?;
-    let hdr_style = Style::new().bold().underline();
-    log::debug!("{}", "Check Push".style(hdr_style));
-    log::debug!("Branch status: {}", client.branch_status()?);
-
-    Ok(())
-}
-
-async fn commit_changed_files(
-    client: &Client,
-    sign: Sign,
-    commit_message: &str,
-    prefix: &str,
-    tag_opt: Option<&str>,
-) -> Result<()> {
-    let hdr_style = Style::new().bold().underline();
-    log::debug!("{}", "Check WorkDir".style(hdr_style));
-
-    let files_in_workdir = client.repo_files_not_staged()?;
-
-    log::debug!("WorkDir files:\n\t{:?}", files_in_workdir);
-    log::debug!("Staged files:\n\t{:?}", client.repo_files_staged()?);
-    log::debug!("Branch status: {}", client.branch_status()?);
-
-    log::info!("Stage the changes for commit");
-
-    client.stage_files(files_in_workdir)?;
-
-    log::debug!("{}", "Check Staged".style(hdr_style));
-    log::debug!("WorkDir files:\n\t{:?}", client.repo_files_not_staged()?);
-
-    let files_staged_for_commit = client.repo_files_staged()?;
-
-    log::debug!("Staged files:\n\t{:?}", files_staged_for_commit);
-    log::debug!("Branch status: {}", client.branch_status()?);
-
-    log::info!("Commit the staged changes");
-
-    client.commit_staged(sign, commit_message, prefix, tag_opt)?;
-
-    log::debug!("{}", "Check Committed".style(hdr_style));
-    log::debug!("WorkDir files:\n\t{:?}", client.repo_files_not_staged()?);
-
-    let files_staged_for_commit = client.repo_files_staged()?;
-
-    log::debug!("Staged files:\n\t{:?}", files_staged_for_commit);
-    log::debug!("Branch status: {}", client.branch_status()?);
-
-    Ok(())
-}
-
-async fn run_release(sign: Sign, args: Release) -> Result<CIExit> {
-    let client = get_client(Commands::Release(args.clone())).await?;
-
-    if args.workspace {
-        log::info!("Running release for workspace");
-        return release_workspace(client, args).await;
-    };
-
-    if args.package.is_some() {
-        return release_package(client, args).await;
-    }
-
-    release_semver(client, args, sign).await
-}
-
-async fn release_semver(mut client: Client, args: Release, sign: Sign) -> Result<CIExit> {
-    let Some(version) = args.semver else {
-        log::error!("Semver required to update changelog");
-        return Ok(CIExit::UnChanged);
-    };
-
-    log::trace!("Running release {version}");
-    log::trace!(
-        "PR ID: {} - Owner: {} - Repo: {}",
-        client.pr_number(),
-        client.owner(),
-        client.repo()
-    );
-    log::trace!("Signing: {:?}", sign);
-    log::trace!("Update changelog flag: {}", args.update_changelog);
-
-    if args.update_changelog {
-        client.release_unreleased(&version)?;
-        log::debug!("Changelog file name: {}", client.changelog_as_str());
-
-        log::trace!(
-            "{}",
-            print_changelog(client.changelog_as_str(), client.line_limit())
-        );
-
-        let commit_message = "chore: update changelog for pr";
-
-        commit_changed_files(&client, sign, commit_message, &args.prefix, Some(&version)).await?;
-
-        push_committed(&client, &args.prefix, Some(&version), false).await?;
-    }
-
-    client.make_release(&args.prefix, &version).await?;
-
-    Ok(CIExit::Released)
-}
-
-async fn release_workspace(client: Client, args: Release) -> Result<CIExit> {
-    let path = Path::new("./Cargo.toml");
-    let workspace = Workspace::new(path).unwrap();
-
-    let packages = workspace.packages();
-
-    if let Some(packages) = packages {
-        for package in packages {
-            let prefix = format!("{}-{}", package.name, args.prefix);
-            let version = package.version;
-            let tag = format!("{prefix}{version}");
-            if !client.tag_exists(&tag) {
-                log::error!("Tag does not exist: {tag}");
-            } else {
-                log::info!("Tag already exists: {tag}, attempt to make release");
-                client.make_release(&prefix, &version).await?;
-            }
-        }
-    }
-    Ok(CIExit::Released)
-}
-
-async fn release_package(client: Client, args: Release) -> Result<CIExit> {
-    let rel_package = args.package.unwrap();
-    log::info!("Running release for package: {}", rel_package);
-
-    let path = Path::new("./Cargo.toml");
-    let workspace = Workspace::new(path).unwrap();
-
-    let packages = workspace.packages();
-
-    if let Some(packages) = packages {
-        for package in packages {
-            log::debug!("Found workspace package: {}", package.name);
-            if package.name != rel_package {
-                continue;
-            }
-            let prefix = format!("{}-{}", package.name, args.prefix);
-            let version = package.version;
-            let tag = format!("{prefix}{version}");
-            if !client.tag_exists(&tag) {
-                log::error!("Tag does not exist: {tag}");
-            } else {
-                log::info!("Tag already exists: {tag}, attempt to make release");
-                client.make_release(&prefix, &version).await?;
-            }
-            break;
-        }
-    }
-    Ok(CIExit::Released)
-}
-
-fn print_changelog(changelog_path: &str, mut line_limit: usize) -> String {
-    let mut output = String::new();
-
-    if let Ok(change_log) = fs::read_to_string(changelog_path) {
-        let mut line_count = 0;
-        if line_limit == 0 {
-            line_limit = change_log.lines().count();
-        };
-
-        output.push_str("\n*****Changelog*****:\n----------------------------");
-        for line in change_log.lines() {
-            output.push_str(format!("{line}\n").as_str());
-            line_count += 1;
-            if line_count >= line_limit {
-                break;
-            }
-        }
-        output.push_str("----------------------------\n");
-    };
-
-    output
-}
-
 fn get_logging(level: &log::LevelFilter) -> env_logger::Builder {
     let env = Env::new()
         .filter_or(LOG_ENV_VAR, "off")
@@ -253,13 +60,6 @@ fn get_logging(level: &log::LevelFilter) -> env_logger::Builder {
     builder.format_timestamp_secs();
 
     builder
-}
-
-async fn get_client(cmd: Commands) -> Result<Client, Error> {
-    let settings = get_settings(cmd)?;
-    let client = Client::new_with(settings).await?;
-
-    Ok(client)
 }
 
 fn get_tracing(level: log::LevelFilter) {
@@ -276,54 +76,4 @@ fn get_tracing(level: log::LevelFilter) {
         .map_err(|_| eprintln!("Unable to set global default subscriber!"));
 
     tracing::info!("Initialised logging to console at {level}");
-}
-
-fn get_settings(cmd: Commands) -> Result<Config, Error> {
-    let mut settings = Config::builder()
-        // Set defaults for CircleCI
-        .set_default("log", "CHANGELOG.md")?
-        .set_default("branch", "CIRCLE_BRANCH")?
-        .set_default("default_branch", "main")?
-        .set_default("pull_request", "CIRCLE_PULL_REQUEST")?
-        .set_default("username", "CIRCLE_PROJECT_USERNAME")?
-        .set_default("reponame", "CIRCLE_PROJECT_REPONAME")?
-        .set_default("commit_message", "chore: update changelog")?
-        .set_default("dev_platform", "https://github.com/")?
-        .set_default("version_prefix", "v")?
-        // Add in settings from pcu.toml if it exists
-        .add_source(config::File::with_name("pcu.toml").required(false))
-        // Add in settings from the environment (with a prefix of PCU)
-        .add_source(config::Environment::with_prefix("PCU"));
-
-    settings = match cmd {
-        Commands::Pr(_) => settings
-            .set_override("commit_message", "chore: update changelog for pr")?
-            .set_override("command", "pr")?,
-        Commands::Release(_) => settings
-            .set_override("commit_message", "chore: update changelog for release")?
-            .set_override("command", "release")?,
-        Commands::Commit(_) => settings
-            .set_override("commit_message", "chore: adding changed files")?
-            .set_override("command", "commit")?,
-        Commands::Push(_) => settings
-            .set_override("commit_message", "chore: update changelog for release")?
-            .set_override("command", "push")?,
-        Commands::Label(_) => settings
-            .set_override("commit_message", "chore: update changelog for release")?
-            .set_override("command", "label")?,
-    };
-
-    settings = if let Ok(pat) = env::var(GITHUB_PAT) {
-        settings.set_override("pat", pat.to_string())?
-    } else {
-        settings
-    };
-
-    match settings.build() {
-        Ok(settings) => Ok(settings),
-        Err(e) => {
-            log::error!("Error: {e}");
-            Err(e.into())
-        }
-    }
 }
